Add CRUD for customer object
Adding CRUD operations for the Customer list in
the TransportCompany Object. Along with that a
workaround had to be though of in order to pass
data around controllers, instead of using a Model.
This is achieved through init method with which the
array of transport companies are passed between
windows and every windows work on it

Signed-off-by: Martin Dekov <mvdekov@gmail.com>